pyproject.toml
[project]
name = "strategy_ai"
version = "0.1.0"
description = "strategy_ai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.9.3",
    "pandas>=2.0",
    "pandas_ta>=0.3.14b1",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "pydantic>=2.0",
]

[project.scripts]
strategy_ai = "strategy_ai.main:run"
run_crew = "strategy_ai.main:run"
train = "strategy_ai.main:train"
replay = "strategy_ai.main:replay"
test = "strategy_ai.main:test"
run_with_trigger = "strategy_ai.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
